server/models/order_tax_test.go
package models

//
//import (
//	"github.com/stretchr/testify/assert"
//	"github.com/vavilen84/nft-project/constants"
//	"github.com/vavilen84/nft-project/validation"
//	"testing"
//)
//
//func TestOrderTax_ValidateOnCreate(t *testing.T) {
//	m := OrderTax{}
//	err := validation.ValidateByScenario(constants.ScenarioCreate, &m, m.getValidator(), m.getValidationRules())
//	assert.NotNil(t, err)
//	assert.NotEmpty(t, err[constants.OrderTaxOrderIdField])
//	assert.NotEmpty(t, err[constants.OrderTaxTaxIdField])
//
//	m = OrderTax{
//		OrderId: 1,
//		TaxId:   1,
//	}
//	err = validation.ValidateByScenario(constants.ScenarioCreate, &m, m.getValidator(), m.getValidationRules())
//	assert.NotNil(t, err)
//}
